requirements.txt
flask==3.1.3
gunicorn==23.0.0
python-dotenv==1.2.1
requests==2.32.5
